Do not cache volume id in variable
The PARTUUID code was originally written for the boot-from-USB
scenario in which only one disk is used.
In that situation it made sense to retrieve the volume id of
the disk once, and cache it in a variable.

However in the boot from SD card, root on USB drive scenario
different drives are used, so this doesn't work.